feat: Updated modal
