+++ src/main/js/scss/__popup-component.scss
@@ -5,7 +5,7 @@
     background-color: rgba(128, 128, 128, 0.493);
     height: 100vh;
     width: 100vw;
-    z-index: 3;
+    z-index: 5;
 
     #popup__container {
         background-color: #ffffff;
@@ -35,7 +35,7 @@
 
         h1 {
             margin: 0;
-            height:fit-content;
+            height: fit-content;
         }
 
         p {
@@ -50,7 +50,6 @@
             align-items: center;
             background-color: #dcecf5;
         }
-
     }
 }
